requirements.txt
behave==1.2.6
behavex==4.1.0
Faker==36.2.2
names==0.3.0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0